fix: rename style folder to styles and resolve duplicates
