--- Frontend/src/styles/mainpage.css
@@ -1,435 +1,0 @@
-.landing-container {
-  min-height: 100vh;
-  background: linear-gradient(135deg, hsl(258, 90%, 66%) 0%, #a855f7 25%, #c084fc 50%, #e879f9 75%, #f472b6 100%);
-  position: relative;
-  overflow-x: hidden;
-}
-
-.landing-header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 1.5rem 2rem;
-  position: relative;
-  z-index: 10;
-}
-
-.logo {
-  display: flex;
-  align-items: center;
-  gap: 0.75rem;
-}
-
-.logo h2 {
-  font-size: 1.5rem;
-  font-weight: bold;
-  color: white;
-  margin: 0;
-}
-
-.logo-icon {
-  width: 3rem;
-  height: 3rem;
-  background: rgba(255, 255, 255, 0.2);
-  border-radius: 50%; 
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  backdrop-filter: blur(10px);
-  border: 1px solid rgba(255, 255, 255, 0.2);
-  overflow: hidden; 
-}
-
-.logo-image {
-  width: 100%;
-  height: 100%;
-  object-fit: cover; 
-  border-radius: 50%; 
-  filter: 
-    brightness(1.3) 
-    contrast(1.2) 
-    drop-shadow(0 0 8px rgba(255, 255, 255, 0.3));
-}
-
-.nav-menu {
-  display: flex;
-  align-items: center;
-  gap: 2rem;
-}
-
-.nav-menu a {
-  color: white;
-  text-decoration: none;
-  transition: color 0.3s ease;
-}
-
-.nav-menu a:hover {
-  color: rgba(255, 255, 255, 0.8);
-}
-
-.nav-buttons {
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-}
-
-.nav-btn {
-  background: none;
-  border: none;
-  color: white;
-  cursor: pointer;
-  padding: 0.5rem 1rem;
-  border-radius: 0.5rem;
-  transition: all 0.3s ease;
-  font-size: 1rem;
-}
-
-.nav-btn.login-btn:hover {
-  color: rgba(255, 255, 255, 0.8);
-}
-
-.nav-btn.signup-btn {
-  background: rgba(255, 255, 255, 0.2);
-  backdrop-filter: blur(10px);
-  border: 1px solid rgba(255, 255, 255, 0.2);
-}
-
-.nav-btn.signup-btn:hover {
-  background: rgba(255, 255, 255, 0.3);
-}
-
-.landing-main {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 4rem 2rem;
-  max-width: 1400px;
-  margin: 0 auto;
-  gap: 4rem;
-}
-
-.hero-section {
-  flex: 1;
-  max-width: 600px;
-}
-
-.hero-title {
-  font-size: 4rem;
-  font-weight: bold;
-  color: white;
-  line-height: 1.1;
-  margin-bottom: 2rem;
-  font-style: italic;
-}
-
-.hero-subtitle {
-  font-size: 1.25rem;
-  color: rgba(255, 255, 255, 0.9);
-  line-height: 1.6;
-  margin-bottom: 3rem;
-  max-width: 500px;
-}
-
-.cta-primary {
-  display: inline-flex;
-  align-items: center;
-  background: rgba(255, 255, 255, 0.2);
-  color: white;
-  padding: 1rem 2rem;
-  border-radius: 9999px;
-  font-size: 1.125rem;
-  font-weight: 600;
-  border: 1px solid rgba(255, 255, 255, 0.2);
-  backdrop-filter: blur(10px);
-  cursor: pointer;
-  transition: all 0.3s ease;
-  text-decoration: none;
-  gap: 0.75rem;
-}
-
-.cta-primary:hover {
-  background: rgba(255, 255, 255, 0.3);
-  transform: translateY(-2px);
-}
-
-.cta-arrow {
-  transition: transform 0.3s ease;
-}
-
-.cta-primary:hover .cta-arrow {
-  transform: translateX(4px);
-}
-
-.illustration-section {
-  flex: 1;
-  display: flex;
-  justify-content: center;
-  position: relative;
-}
-
-.illustration-container {
-  background: rgba(255, 255, 255, 0.1);
-  backdrop-filter: blur(16px);
-  border-radius: 1.5rem;
-  padding: 2rem;
-  border: 1px solid rgba(255, 255, 255, 0.2);
-  max-width: 500px;
-  width: 100%;
-  position: relative;
-}
-
-.mock-dashboard {
-  background: white;
-  border-radius: 1rem;
-  padding: 1.5rem;
-  box-shadow: 0 25px 50px -12px rgba(0, 0, 0, 0.25);
-  position: relative;
-}
-
-.dashboard-header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 1.5rem;
-}
-
-.window-controls {
-  display: flex;
-  gap: 0.5rem;
-}
-
-.window-dot {
-  width: 0.75rem;
-  height: 0.75rem;
-  border-radius: 50%;
-}
-
-.dot-red { background-color: #ef4444; }
-.dot-yellow { background-color: #eab308; }
-.dot-green { background-color: #22c55e; }
-
-.dashboard-amount {
-  font-size: 0.875rem;
-  color: #6b7280;
-  font-weight: 600;
-}
-
-.dashboard-image-area {
-  background: linear-gradient(135deg, #f3e8ff 0%, #dbeafe 100%);
-  border-radius: 0.75rem;
-  padding: 1rem;
-  margin-bottom: 1.5rem;
-  position: relative;
-  overflow: hidden;
-}
-
-.dashboard-preview-image {
-  width: 100%;
-  height: auto;
-  max-height: 200px;
-  object-fit: cover;
-  border-radius: 0.5rem;
-  box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1);
-  transition: transform 0.3s ease;
-}
-
-.dashboard-preview-image:hover {
-  transform: scale(1.02);
-}
-
-.people-section {
-  display: flex;
-  justify-content: space-around;
-  align-items: flex-end;
-}
-
-.person {
-  text-align: center;
-}
-
-.person-avatar {
-  width: 3rem;
-  height: 4rem;
-  border-radius: 1.5rem 1.5rem 0.5rem 0.5rem;
-  margin-bottom: 0.5rem;
-  position: relative;
-}
-
-.person-head {
-  width: 1.5rem;
-  height: 1.5rem;
-  border-radius: 50%;
-  margin: 0.5rem auto 0;
-}
-
-.person-1 {
-  background: linear-gradient(to bottom, #a78bfa, #8b5cf6);
-}
-
-.person-1 .person-head {
-  background: #c4b5fd;
-}
-
-.person-2 {
-  background: linear-gradient(to bottom, #60a5fa, #3b82f6);
-}
-
-.person-2 .person-head {
-  background: #93c5fd;
-}
-
-.person-3 {
-  background: linear-gradient(to bottom, #818cf8, #6366f1);
-}
-
-.person-3 .person-head {
-  background: #a5b4fc;
-}
-
-.plant-1 {
-  position: absolute;
-  bottom: -1.5rem;
-  left: -1.5rem;
-  width: 4rem;
-  height: 5rem;
-  background: linear-gradient(to top, #22c55e, #4ade80);
-  border-radius: 2rem 2rem 0 0;
-  transform: rotate(12deg);
-  opacity: 0.8;
-}
-
-.plant-2 {
-  position: absolute;
-  top: -1.5rem;
-  right: -1.5rem;
-  width: 3rem;
-  height: 4rem;
-  background: linear-gradient(to top, #16a34a, #22c55e);
-  border-radius: 1.5rem 1.5rem 0 0;
-  transform: rotate(-12deg);
-  opacity: 0.8;
-}
-
-.floating-element {
-  position: absolute;
-  border-radius: 50%;
-  animation: float 3s ease-in-out infinite;
-}
-
-.float-1 {
-  top: 25%;
-  left: 25%;
-  width: 1rem;
-  height: 1rem;
-  background: rgba(255, 255, 255, 0.2);
-}
-
-.float-2 {
-  top: 33%;
-  right: 25%;
-  width: 1.5rem;
-  height: 1.5rem;
-  background: rgba(255, 255, 255, 0.1);
-  animation-delay: -1s;
-}
-
-.float-3 {
-  bottom: 25%;
-  left: 33%;
-  width: 0.75rem;
-  height: 0.75rem;
-  background: rgba(255, 255, 255, 0.3);
-  animation-delay: -0.5s;
-}
-
-@keyframes float {
-  0%, 100% {
-    transform: translateY(0px);
-    opacity: 0.7;
-  }
-  50% {
-    transform: translateY(-10px);
-    opacity: 1;
-  }
-}
-
-.mobile-menu-btn {
-  display: none;
-  position: fixed;
-  top: 1.5rem;
-  right: 1.5rem;
-  background: rgba(255, 255, 255, 0.2);
-  padding: 0.5rem;
-  border-radius: 0.5rem;
-  backdrop-filter: blur(10px);
-  border: none;
-  cursor: pointer;
-  z-index: 20;
-}
-
-@media (max-width: 1024px) {
-  .hero-title {
-    font-size: 3rem;
-  }
-  
-  .landing-main {
-    flex-direction: column;
-    text-align: center;
-    gap: 3rem;
-  }
-  
-  .illustration-container {
-    max-width: 400px;
-  }
-}
-
-@media (max-width: 768px) {
-  .nav-menu {
-    display: none;
-  }
-  
-  .mobile-menu-btn {
-    display: block;
-  }
-  
-  .landing-header {
-    padding: 1rem;
-  }
-  
-  .landing-main {
-    padding: 2rem 1rem;
-  }
-  
-  .hero-title {
-    font-size: 2.5rem;
-  }
-  
-  .hero-subtitle {
-    font-size: 1.125rem;
-  }
-  
-  .illustration-container {
-    padding: 1.5rem;
-    max-width: 350px;
-  }
-}
-
-@media (max-width: 640px) {
-  .hero-title {
-    font-size: 2rem;
-  }
-  
-  .nav-buttons {
-    gap: 0.5rem;
-  }
-  
-  .nav-btn {
-    padding: 0.5rem 0.75rem;
-    font-size: 0.875rem;
-  }
-  
-  .cta-primary {
-    padding: 0.875rem 1.5rem;
-    font-size: 1rem;
-  }
-}
